udpate proposed changes query
