+++ frontend/app/src/pages/proposed-changes/proposed-changes-details.tsx
@@ -2,7 +2,7 @@ import { Tabs } from "@/components/tabs";
 import { Link } from "@/components/ui/link";
 import { PROPOSED_CHANGES_OBJECT, TASK_OBJECT, TASK_TAB } from "@/config/constants";
 import { QSP } from "@/config/qsp";
-import { getProposedChanges } from "@/graphql/queries/proposed-changes/getProposedChanges";
+import { getProposedChanges } from "@/graphql/queries/proposed-changes/getProposedChangesDetails";
 import useQuery from "@/hooks/useQuery";
 import { useTitle } from "@/hooks/useTitle";
 import { ArtifactsDiff } from "@/screens/diff/artifact-diff/artifacts-diff";
